Tank (queen) has been added + handling/improvements
If you send a unit to a tile, that tile will be marked as taken the moment you send the unit there. It will no longer wait for the unit to arrive. This allows multiple units to travel at once without actually some of those units taking/going through a tile that will be blocked in the current turn